ARM: dts: add OLPC XO 1.75 machine
This is a fairly complete description of an OLPC XO 1.75 laptop.

It's notably missing:

* LCDC. I didn't manage to get the Armada DRM driver working yet.
* Vivante GPU. At least the device tree part should be straightforward
  once the LCDC works.
* Audio.

I'll add those as time permits.

The laptop's native OpenFirmware is actually able to flatten the device
tree and pass it to the kernel. However, even with the most recent
firmware images, there are significant differencies between the bindings
that firmware provides and the bindings that drivers actually expect.

Given the firmware is actually free software it might be possible to
fix it. Alternatively, it can just be patched around with a Forth boot
script.

So far, I've tested this with CONFIG_ARM_APPENDED_DTB.

Signed-off-by: Lubomir Rintel <lkundrak@v3.sk>